==> code /02_SG_visit_data.ipynb <==
This python scripts is for calculating county level visit rate and brand accessibility.

# packages


import json
import pandas as pd
import os
#!pip install geopandas
import geopandas as gpd
#!pip install -q --upgrade git+https://github.com/SafeGraphInc/safegraph_py
from safegraph_py_functions import safegraph_py_functions as sgpy


os.chdir('/your/path')


# read raw data for POIs of clothing stores

You need to have raw visit data stored in the path, the data used in this study was accessed from Dewey (https://www.deweydata.io/).
The data structure and variable names may have changed. Please verify before using the code below.



The following function reads all raw visit data files. The raw data is on poi (points of interest, i.e., a location) level, with a json column named "visitor_home_cbgs" storing visitor information.

def read_raw(month):
    df_poi_monthly = pd.DataFrame()
  # there are 30 sub files for each month
    for i in range(1, 30):
        file_name = f"/yourpath/raw_data/safegraph_{month}_part{i}.csv.gz"
        df_part = pd.read_csv(file_name,compression='zip',usecols=[0,4,6,7,9,10,28,29,30,35,36]) #select essential columns 
        # filter data by poi category and country
        df_part = df_part[(df_part.top_category == "Clothing Stores") & (df_part.iso_country_code == "US")]
        df_part['visitor_home_cbgs'] = df_part['visitor_home_cbgs'].fillna('')
        # remove pois without visitor information
        df_part = df_part[(df_part['visitor_home_cbgs'] != '') & (df_part['visitor_home_cbgs'] != '{}')]
        # combine 30 files of the same month
        df_poi_monthly = pd.concat([df_poi_monthly, df_part], axis=0)
        # add month column
        df_poi_monthly['month'] = month
        return(df_poi_monthly)

# read the data for each month and combine
dfs={}
all_months = ['jan','feb','mar','apr','may','jun','jul','aug','sep','oct','nov','dec']
dfs={}
for month in all_months:
    df_poi_monthly = read_raw(month)
    dfs[f'df_poi_{month}'] = df_poi_monthly

df_poi_raw = pd.concat(dfs.values(), ignore_index=True)

# assign brand tiers

# survey_ratings.csv is produced from 01_brand_ratings.R
survey_data = pd.read_csv('clean_data/survey_ratings.csv')

survey_data = survey_data[~survey_data.brands.isin(['Nordstrom','Marshalls','Burlington'])]#remove department stores
# 100 most recognized brands
recognizable_brands = survey_data.sort_values('perc_na')[:100]
# sort by average ratings
recognizable_brands = recognizable_brands.sort_values(by='mean',ascending=False)
# brand lists of 4 tiers
tier1_brands = recognizable_brands.iloc[0:25,]['brands'].values
tier2_brands = recognizable_brands.iloc[25:50,]['brands'].values
tier3_brands = recognizable_brands.iloc[50:75,]['brands'].values
tier4_brands = recognizable_brands.iloc[75:100,]['brands'].values

# assign tier to brands
def label_tier(brand):
    if pd.isnull(brand):
        return('no_brands')
    elif brand in tier1_brands:
        return('tier1')
    elif brand in tier2_brands:
        return('tier2')
    elif brand in tier3_brands:
        return('tier3')
    elif brand in tier4_brands:
        return('tier4')
    else: return('other_brands')

def preprocess_poi(df_poi):
    df_poi['brand_tier'] = df_poi.brands.apply(lambda x: label_tier(x)) # add brand tier column  
    df_poi = df_poi[df_poi.brand_tier != "no_brands"]# remove poi with no brand recorded in the data
    df_poi['poi_cbg'] = df_poi['poi_cbg'].astype('str')  
    # remove US outlying territories
    df_poi = df_poi.drop(df_poi[df_poi.poi_cbg.str.startswith(('60','66','69','78','72'))].index).reset_index(drop=True)
    return(df_poi)

df_poi = preprocess_poi(df_poi_raw)

# county level visit rate

we calculate number of 2 types of visitors for each county:
1.   visitors who visited all clothing stores
2.   visitors who visited a certain brand tier



# Function to calculate the monthly average number of visitors for all POIs
# or for a specific tier (if a tier is specified).

def get_visitors(df_poi,tier):
    # filter POI data for the specified brand tier (if provided)
    if tier is not None:
        df_poi = df_poi[df_poi.brand_tier == tier]
    
    # Unpack json column
    df_home = sgpy.unpack_json_fast(df_poi, chunk_n=1000) # chunk_n is the size of one chunk
    
    # get county fips
    df_home['home_county'] = df_home.visitor_home_cbgs_key.str[:5]
    
    # calculate county level visitors
    df_county_home = df_home.groupby('home_county')['visitor_home_cbgs_value'].sum().reset_index()
    df_county_home = df_county_home.rename(columns = {'visitor_home_cbgs_value': 'visitors'})
    
    # calculate monthly average number of visitors
    df_county_home['visitors'] = df_county_home['visitors']/len(df_poi.month.unique()) #calculate average monthly visitors
    
    # Drop counties in oversea territories and whose FIPS code starts with "CA"
    df_county_home = df_county_home[~df_county_home.home_county.str.startswith(('60', '66', '69', '78', '72', 'CA'))].reset_index(drop=True)
    
    # add a tier column
    df_county_home['tier'] = tier if tier is not None else "all"
    
    return df_county_home

# get county level visitors for all pois 
df_all_visitors = get_visitors(df_poi,tier=None)
df_all_visitors = df_all_visitors.rename(columns = {"visitors" : "all_visitors"})

# calculate county visitors for each tier

all_tiers = ['tier1','tier2','tier3','tier4']
dfs={}

for tier in all_tiers:
    df_home = get_tier_visitors(df_poi,tier)
    dfs[f"df_{tier}_visitors"] = df_home
    
df_tier_visitors = pd.concat(dfs.values(), ignore_index=True)
# convert data from long to wide format 
df_tier_visitors_wide=df_tier_visitors.pivot(index='home_county',columns='tier',values="tier_visitors").fillna(0).reset_index()


# merge tier visitors with all visitors
df_visitors = df_all_visitors.merge(df_tier_visitors_wide,on="home_county",how='left').fillna(0)
# wide to long format
df_visitors_long = pd.melt(df_visitors, id_vars=['home_county','all_visitors'],
                           value_vars=['tier1','tier2','tier3','tier4'],
                           var_name='tier', value_name='tier_visitors')
# merging index will be used for merging with store accessibility data
df_visitors_long['merge_index'] = df_visitors_long.apply(lambda x: x.home_county+x.tier,axis=1)

# brand accessibility

We calculate the percent of brand tier that are accessible to each home county. we expanded the polygon of each county by 20 miles beyond its original boundary.  Therefore, not only are these POIs located within the county boundary, but also those located near enough are considered accessible.

# Load the json file with county coordinates
geoData = gpd.read_file('https://raw.githubusercontent.com/holtzy/The-Python-Graph-Gallery/master/static/data/US-counties.geojson')


def poi_accessibility(df_poi):
    # select poi cordinates
    df_poi2=df_poi.loc[:,['latitude','longitude','brand_tier']]
    #transform points data to geodf format
    geodf_poi = gpd.GeoDataFrame(df_poi2, geometry=gpd.points_from_xy(df_poi2.longitude, df_poi2.latitude),crs='EPSG:4326')
    #transform points data to epsg27700 format (measured in meters)
    geodf_poi2=geodf_poi.to_crs('EPSG:27700')
    #expand county boundaries by 20 miles
    buffers_20mile = geoData.set_geometry(geoData.to_crs('EPSG:27700').buffer(32187))
    #merge county polygon and poi points
    dfmerge_county = gpd.sjoin(buffers_20mile, geodf_poi2, how="left", predicate='intersects')
    # calculate number of pois of different tiers within each county, "id" represents county code
    dfmerge_county = dfmerge_county[['id','brand_tier']].groupby(['id','brand_tier']).size().reset_index()
    dfmerge_county = dfmerge_county.rename(columns={0:'counts'}) #remame column
    #average monthly counts, as df_poi includes 12 month data
    dfmerge_county['counts']=dfmerge_county['counts']/len(df_poi.month.unique())
    #count points of all brand type
    total_counts = dfmerge_county.groupby('id')['counts'].sum()
    # number of total accessible clothing stores
    dfmerge_county['total_counts'] = dfmerge_county.id.apply(lambda x: total_counts[x])
    # percent of accessible stores of given brand tier
    dfmerge_county['store_ratio'] = dfmerge_county.apply(lambda x: x.counts/x.total_counts,axis=1)
    # merging index that will be used for merging with visitor data
    dfmerge_county['merge_index'] = dfmerge_county.apply(lambda x: x.id+x.brand_tier,axis=1)

    return(dfmerge_county)

df_brand_accessibility = poi_accessibility(df_poi)

# merge visitor data with brand accessibility data
df_visitors_acc = df_visitors_long.merge(df_brand_accessibility[['merge_index','store_ratio']],
                                         on = 'merge_index',
                                         how='left').fillna(0)

df_visitors_acc.to_csv("/clean_data/df_visits.csv",index=False)
